requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==2.0.12
cryptography==43.0.1
defusedxml==0.7.1
dj-database-url==1.2.0
dj-rest-auth==3.0.0
Django==5.1.1
django-allauth==0.52.0
django-simpleui==2024.8.28
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
gunicorn==20.1.0
idna==3.8
multidict==6.1.0
oauthlib==3.2.2
openai==0.27.8
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2024.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
soupsieve==2.6
sqlparse==0.5.1
tiktoken==0.4.0
tqdm==4.66.5
urllib3==1.26.20
yarl==1.11.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
pydantic==2.9.1
